refactor: keep global variables in a dedicated folder
