+++ src/components/ItemList/ItemList.tsx
@@ -5,8 +5,8 @@ import Paginator from '../Paginator/Paginator';
 import Loader from '../Loader/Loader';
 import { queryItems } from '../API/API';
 import { EMPTY_ITEMS_ARRAY } from './data';
-import { SEARCH_DEFAULT } from '../../data';
-import { MAX_AMOUNT } from '../../data';
+import { SEARCH_DEFAULT } from '../../shared/data/data';
+import { MAX_AMOUNT } from '../../shared/data/data';
 
 export const DataFromChildContext = createContext<React.Dispatch<
   React.SetStateAction<boolean>
